source-application: startup error handling
When the streaming pipeline fails to build, exit the application
immediately with nonzero status. This allows the init system or
higher-level scripts to detect there was a problem with starting the
source.

The initialization code got moved into GApplication::command_line, which
can make g_application_run() quit and return a status code of our
choice.