source-application: fix exposing a channel on the system bus
Since GApplication doesn't support automatic registering on the system
bus, GaeulApplication calls g_bus_own_name() to do so on its own. The
call is, however, asynchronous and might not finish before we want to
export a channel on the acquired D-Bus connection in
gaeul_source_application_command_line().

To overcome that, don't rely on the application's dbus_service and
acquire the connection for exposing channels synchronously based on the
value of "dbus-type" property.